refactor: move test utils to separate file
